fix(ui): keep model picker search current across refresh and Escape
