+++ ui/src/e2e/chat-flow.model-picker-refresh.e2e.test.ts
@@ -2,6 +2,7 @@ import { mkdir } from "node:fs/promises";
 import path from "node:path";
 import type { Page } from "playwright";
 import { expect, it } from "vitest";
+import { createControlUiE2eArtifactDir } from "../test-helpers/control-ui-e2e-artifacts.ts";
 import {
   createChatFlowE2eSuite,
   installMockGateway,
@@ -301,6 +302,95 @@ suite.define(() => {
         .waitFor({ state: "visible" });
       expect(await picker.locator("[data-chat-model-catalog-state]").count()).toBe(0);
       await screenshot(page, "02-picker-after-background-apply.png");
+    } finally {
+      await context.close();
+    }
+  });
+
+  it("reconciles the current model search when an open catalog replaces its results", async () => {
+    const context = await suite.newBrowserContext(createControlUiE2eContextOptions());
+    const page = await context.newPage();
+    const artifactRoot = process.env.OPENCLAW_UI_E2E_ARTIFACT_DIR?.trim();
+    const artifactDir = artifactRoot
+      ? createControlUiE2eArtifactDir("model-search-refresh", artifactRoot)
+      : undefined;
+    const gateway = await installMockGateway(page, {
+      models: [
+        { id: "gpt-5.5", name: "GPT-5.5", provider: "openai" },
+        { id: "claude-haiku-4-5", name: "Claude Haiku 4.5", provider: "anthropic" },
+      ],
+      sessionKey: "agent:main:main",
+    });
+    try {
+      await page.goto(`${suite.server.baseUrl}chat`);
+      const picker = page.locator(
+        'openclaw-chat-pane[aria-hidden="false"] .chat-controls__model-picker',
+      );
+      const previous = picker.locator('[data-chat-model-option="anthropic/claude-haiku-4-5"]');
+      await previous.waitFor({ state: "attached" });
+      const discoveryCount = (await gateway.getRequests("models.list")).length;
+      await gateway.deferNext("models.list", { refresh: true });
+      await picker.locator('[data-chat-model-select="true"]').click();
+      await gateway.waitForRequest("models.list", { after: discoveryCount });
+      const search = picker.locator("[data-chat-model-search]");
+      await search.fill("anthropic");
+      await expect.poll(() => picker.locator("[data-chat-model-option]:visible").count()).toBe(1);
+      expect(await previous.isVisible()).toBe(true);
+      if (artifactDir) {
+        await page.screenshot({
+          animations: "disabled",
+          path: `${artifactDir}/01-search-warm.png`,
+        });
+      }
+
+      const metadataCount = (await gateway.getRequests("chat.metadata")).length;
+      await gateway.deferNext("chat.metadata");
+      const refreshedModels = [
+        { id: "gpt-5.5", name: "GPT-5.5", provider: "openai" },
+        { id: "gpt-5.4-mini", name: "GPT-5.4 mini", provider: "openai" },
+        { id: "claude-sonnet-4-6", name: "Claude Sonnet 4.6", provider: "anthropic" },
+      ];
+      await gateway.resolveDeferred("models.list", { models: refreshedModels });
+      await gateway.waitForRequest("chat.metadata", { after: metadataCount });
+      expect(await previous.isVisible()).toBe(true);
+      await gateway.resolveDeferred("chat.metadata", { commands: [], models: refreshedModels });
+      const replacement = picker.locator('[data-chat-model-option="anthropic/claude-sonnet-4-6"]');
+      await replacement.waitFor({ state: "attached" });
+      await previous.waitFor({ state: "detached" });
+      if (artifactDir) {
+        await page.screenshot({
+          animations: "disabled",
+          path: `${artifactDir}/02-search-refreshed.png`,
+        });
+      }
+
+      expect(await search.inputValue()).toBe("anthropic");
+      await expect
+        .poll(() =>
+          picker
+            .locator("[data-chat-model-option]:visible")
+            .evaluateAll((rows) => rows.map((row) => row.getAttribute("data-chat-model-option"))),
+        )
+        .toEqual(["anthropic/claude-sonnet-4-6"]);
+      expect(await picker.locator("[data-chat-model-search-empty]").isVisible()).toBe(false);
+      await expect
+        .poll(() =>
+          search.evaluate((input) => {
+            const activeId = input.getAttribute("aria-activedescendant");
+            return activeId
+              ? document.getElementById(activeId)?.getAttribute("data-chat-model-option")
+              : null;
+          }),
+        )
+        .toBe("anthropic/claude-sonnet-4-6");
+      const patchCount = (await gateway.getRequests("sessions.patch")).length;
+      await search.press("Enter");
+      const patch = await gateway.waitForRequest("sessions.patch", { after: patchCount });
+      expect(patch.params).toMatchObject({
+        key: "agent:main:main",
+        model: "anthropic/claude-sonnet-4-6",
+      });
+      await expect.poll(() => picker.getAttribute("open")).toBe(null);
     } finally {
       await context.close();
     }
