+++ ui/src/e2e/chat-model-controls.e2e.test.ts
@@ -537,4 +537,61 @@ suite.define(() => {
       });
     },
   );
+
+  it.each(["chat", "new"])(
+    "clears %s model search before Escape dismisses the picker",
+    async (route) => {
+      await suite.withPage({ viewport: { width: 1280, height: 900 } }, async ({ page }) => {
+        const artifactRoot = process.env.OPENCLAW_UI_E2E_ARTIFACT_DIR?.trim();
+        const artifactDir = artifactRoot
+          ? createControlUiE2eArtifactDir(`model-search-escape-${route}`, artifactRoot)
+          : undefined;
+        const gateway = await installMockGateway(page, {
+          agentModel: "openai/gpt-5.5",
+          models: [
+            { id: "gpt-5.5", name: "GPT-5.5", provider: "openai" },
+            { id: "claude-sonnet-4-6", name: "Claude Sonnet 4.6", provider: "anthropic" },
+          ],
+        });
+        await page.goto(`${suite.server.baseUrl}${route}`);
+        const composer = page.locator(".agent-chat__input").first();
+        const picker = composer.locator(".chat-controls__model-picker");
+        const trigger = picker.locator('[data-chat-model-select="true"]');
+        await expect.poll(() => picker.locator("[data-chat-model-option]").count()).toBe(2);
+        await expect.poll(() => trigger.getAttribute("aria-disabled")).toBe("false");
+        await trigger.click();
+        const search = picker.locator("[data-chat-model-search]");
+        await search.fill("anthropic");
+        await expect.poll(() => picker.locator("[data-chat-model-option]:visible").count()).toBe(1);
+        if (artifactDir) {
+          await page.screenshot({
+            animations: "disabled",
+            path: `${artifactDir}/01-filtered.png`,
+          });
+        }
+        await search.press("Escape");
+        if (artifactDir) {
+          await page.screenshot({
+            animations: "disabled",
+            path: `${artifactDir}/02-first-escape.png`,
+          });
+        }
+        await expect.poll(() => picker.getAttribute("open")).toBe("");
+        expect(await search.inputValue()).toBe("");
+        expect(await search.evaluate((input) => input === document.activeElement)).toBe(true);
+        await expect.poll(() => picker.locator("[data-chat-model-option]:visible").count()).toBe(2);
+        expect(await gateway.getRequests("sessions.patch")).toEqual([]);
+
+        await search.press("Escape");
+        await expect.poll(() => picker.getAttribute("open")).toBe(null);
+        expect(await trigger.evaluate((summary) => summary === document.activeElement)).toBe(true);
+        if (artifactDir) {
+          await page.screenshot({
+            animations: "disabled",
+            path: `${artifactDir}/03-second-escape.png`,
+          });
+        }
+      });
+    },
+  );
 });
