+++ ui/src/pages/chat/components/chat-model-picker.ts
@@ -1,4 +1,5 @@
 import { html, nothing } from "lit";
+import { ref } from "lit/directives/ref.js";
 import { repeat } from "lit/directives/repeat.js";
 import { icons } from "../../../components/icons.ts";
 import "../../../components/tooltip.ts";
@@ -8,7 +9,6 @@ import {
   renderProviderBrandIcon,
 } from "../../../components/provider-icon.ts";
 import { t } from "../../../i18n/index.ts";
-import { generateUUID } from "../../../lib/uuid.ts";
 import {
   type ChatContextWindowControlParams,
   renderContextWindowControl,
@@ -24,6 +24,15 @@ import {
   type ChatModelPickerOption,
   type ChatModelPickerTargetGroup,
 } from "./chat-model-picker-options.ts";
+import {
+  handleModelPickerKeydown,
+  handleModelSearchKeydown,
+  highlightModelRow,
+  pickerMenu,
+  resetModelSearch,
+  syncChatModelSearch,
+  updateModelSearch,
+} from "./chat-model-picker-search.ts";
 import { handleChatComposerDetailsToggle, syncChatPickerOverlay } from "./chat-picker-overlay.ts";
 
 export type { ChatModelCatalogState } from "./chat-model-catalog-state.ts";
@@ -56,207 +65,6 @@ type ChatModelPickerParams = {
   onTargetSelect?: (groupId: string, value: string) => unknown;
   onRequestUpdate?: () => void;
 };
-
-function pickerMenu(target: EventTarget | null): HTMLElement | null {
-  return target instanceof Element
-    ? target.closest<HTMLElement>(".chat-controls__model-menu")
-    : null;
-}
-
-function visibleModelRows(root: HTMLElement): HTMLButtonElement[] {
-  return [...root.querySelectorAll<HTMLButtonElement>("[data-chat-model-option]")]
-    .filter((row) => !row.hidden)
-    .toSorted(
-      (left, right) =>
-        Number(left.dataset.chatModelRank ?? left.dataset.chatModelIndex ?? 0) -
-        Number(right.dataset.chatModelRank ?? right.dataset.chatModelIndex ?? 0),
-    );
-}
-
-function selectableModelRows(root: HTMLElement): HTMLButtonElement[] {
-  return visibleModelRows(root).filter((row) => !row.disabled);
-}
-
-function ensureModelPickerIds(menu: HTMLElement): void {
-  const details = menu.closest<HTMLDetailsElement>(".chat-controls__model-picker");
-  const input = menu.querySelector<HTMLInputElement>("[data-chat-model-search]");
-  const listboxes = [...menu.querySelectorAll<HTMLElement>("[data-chat-model-list]")];
-  if (!details || !input || listboxes.length === 0) {
-    return;
-  }
-  const prefix = details.dataset.chatModelPickerId ?? `chat-model-picker-${generateUUID()}`;
-  details.dataset.chatModelPickerId = prefix;
-  listboxes.forEach((listbox, index) => {
-    listbox.id = `${prefix}-listbox-${index}`;
-  });
-  menu.querySelectorAll<HTMLButtonElement>("[data-chat-model-option]").forEach((row, index) => {
-    row.id = `${prefix}-option-${index}`;
-  });
-  input.setAttribute("aria-controls", listboxes.map((listbox) => listbox.id).join(" "));
-  input.setAttribute("aria-expanded", details.open ? "true" : "false");
-}
-
-function highlightModelRow(menu: HTMLElement, row: HTMLButtonElement | undefined): void {
-  menu.querySelectorAll<HTMLElement>("[data-chat-model-option]").forEach((candidate) => {
-    candidate.toggleAttribute("data-chat-model-highlighted", candidate === row);
-  });
-  const input = menu.querySelector<HTMLInputElement>("[data-chat-model-search]");
-  if (row?.id) {
-    input?.setAttribute("aria-activedescendant", row.id);
-  } else {
-    input?.removeAttribute("aria-activedescendant");
-  }
-}
-
-// Numbers follow the filtered order because digit selection reads that same row list.
-// Search inputs and nested dropdowns own digits instead. The :focus-within rule
-// in styles/chat/composer.css hides these keycaps while the search input has focus.
-function updateModelShortcuts(menu: HTMLElement, rows: readonly HTMLButtonElement[]): void {
-  menu.querySelectorAll<HTMLElement>("[data-chat-model-shortcut]").forEach((shortcut) => {
-    shortcut.hidden = true;
-    shortcut.removeAttribute("data-shortcut");
-    shortcut.removeAttribute("data-chat-model-shortcut-number");
-  });
-  rows.slice(0, 9).forEach((row, index) => {
-    const shortcut = row.querySelector<HTMLElement>("[data-chat-model-shortcut]");
-    if (!shortcut) {
-      return;
-    }
-    shortcut.hidden = false;
-    shortcut.setAttribute("data-shortcut", String(index + 1));
-    shortcut.setAttribute("data-chat-model-shortcut-number", String(index + 1));
-  });
-}
-
-function modelMatchRank(row: HTMLButtonElement, query: string): number | null {
-  const model = row.dataset.chatModelName ?? "";
-  const keywords = row.dataset.chatModelKeywords ?? "";
-  const provider = row.dataset.chatModelProviderLabel ?? "";
-  if (model.startsWith(query)) {
-    return 0;
-  }
-  if (model.includes(query)) {
-    return 1;
-  }
-  if (keywords.includes(query)) {
-    return 2;
-  }
-  if (provider.startsWith(query)) {
-    return 3;
-  }
-  return provider.includes(query) ? 4 : null;
-}
-
-function updateModelSearch(input: HTMLInputElement): void {
-  const menu = pickerMenu(input);
-  if (!menu) {
-    return;
-  }
-  ensureModelPickerIds(menu);
-  const query = input.value.trim().toLocaleLowerCase();
-  menu.toggleAttribute("data-chat-model-filtering", Boolean(query));
-  const rows = [...menu.querySelectorAll<HTMLButtonElement>("[data-chat-model-option]")];
-  const matches: Array<{ row: HTMLButtonElement; score: number; index: number }> = [];
-  rows.forEach((row, index) => {
-    const score = query ? modelMatchRank(row, query) : 0;
-    row.hidden = score === null;
-    row.style.removeProperty("--chat-model-rank");
-    delete row.dataset.chatModelRank;
-    if (score !== null) {
-      matches.push({ row, score, index });
-    }
-  });
-  matches
-    .toSorted((left, right) => left.score - right.score || left.index - right.index)
-    .forEach(({ row }, rank) => {
-      row.dataset.chatModelRank = String(rank);
-      row.style.setProperty("--chat-model-rank", String(rank));
-    });
-  const visibleRows = visibleModelRows(menu);
-  const selectableRows = selectableModelRows(menu);
-  updateModelShortcuts(menu, selectableRows);
-  const selected = selectableRows.find((row) => row.getAttribute("aria-selected") === "true");
-  highlightModelRow(menu, query ? selectableRows[0] : (selected ?? selectableRows[0]));
-  const empty = menu.querySelector<HTMLElement>("[data-chat-model-search-empty]");
-  if (empty) {
-    empty.hidden = !query || visibleRows.length > 0;
-  }
-}
-
-function resetModelSearch(details: HTMLDetailsElement): void {
-  const input = details.querySelector<HTMLInputElement>("[data-chat-model-search]");
-  if (!input) {
-    return;
-  }
-  input.value = "";
-  updateModelSearch(input);
-}
-
-export function clearChatModelSearchOnEscape(event: KeyboardEvent): boolean {
-  if (event.key !== "Escape") {
-    return false;
-  }
-  const input = event
-    .composedPath()
-    .find(
-      (target): target is HTMLInputElement =>
-        target instanceof HTMLInputElement && target.matches("[data-chat-model-search]"),
-    );
-  if (!input?.value) {
-    return false;
-  }
-  input.value = "";
-  updateModelSearch(input);
-  event.preventDefault();
-  event.stopPropagation();
-  return true;
-}
-
-function handleModelSearchKeydown(event: KeyboardEvent): void {
-  const input = event.currentTarget as HTMLInputElement;
-  const menu = pickerMenu(input);
-  if (!menu) {
-    return;
-  }
-  const rows = selectableModelRows(menu);
-  if (rows.length === 0) {
-    return;
-  }
-  if (event.key === "Enter") {
-    const highlighted = rows.find((row) => row.hasAttribute("data-chat-model-highlighted"));
-    if (highlighted) {
-      event.preventDefault();
-      highlighted.click();
-    }
-    return;
-  }
-  if (event.key !== "ArrowDown" && event.key !== "ArrowUp") {
-    return;
-  }
-  event.preventDefault();
-  const currentIndex = rows.findIndex((row) => row.hasAttribute("data-chat-model-highlighted"));
-  const offset = event.key === "ArrowDown" ? 1 : rows.length - 1;
-  const nextIndex = currentIndex < 0 ? 0 : (currentIndex + offset) % rows.length;
-  highlightModelRow(menu, rows[nextIndex]);
-  rows[nextIndex]?.scrollIntoView?.({ block: "nearest" });
-}
-
-function handleModelPickerKeydown(event: KeyboardEvent): void {
-  const details = event.currentTarget as HTMLDetailsElement;
-  if (
-    !details.open ||
-    event.target instanceof HTMLInputElement ||
-    event
-      .composedPath()
-      .some((target) => target instanceof HTMLElement && target.localName === "wa-dropdown") ||
-    !/^[1-9]$/u.test(event.key)
-  ) {
-    return;
-  }
-  const row = selectableModelRows(details)[Number(event.key) - 1];
-  event.preventDefault();
-  row?.click();
-}
 
 export function renderChatModelPicker(params: ChatModelPickerParams) {
   const defaultModelOption = params.modelOptions.find((option) => option.isDefault);
@@ -369,6 +177,7 @@ export function renderChatModelPicker(params: ChatModelPickerParams) {
   };
   return html`
     <details
+      ${ref((details) => syncChatModelSearch(details))}
       class="chat-controls__inline-select chat-controls__model-picker"
       data-chat-autotype-shortcuts
       ?open=${params.open === true}
@@ -383,12 +192,7 @@ export function renderChatModelPicker(params: ChatModelPickerParams) {
           return;
         }
         void params.onOpen?.();
-        queueMicrotask(() => {
-          const input = details.querySelector<HTMLInputElement>("[data-chat-model-search]");
-          if (input) {
-            updateModelSearch(input);
-          }
-        });
+        syncChatModelSearch(details);
       }}
     >
       <summary
