+++ ui/src/pages/chat/components/chat-picker-overlay.ts
@@ -1,5 +1,6 @@
 import { syncAnchoredOverlay } from "../../../components/anchored-overlay.ts";
 import { consumeTooltipEscape } from "../../../components/tooltip.ts";
+import { clearChatModelSearchOnEscape } from "./chat-model-picker-search.ts";
 
 const MOBILE_COMPOSER_OVERLAY_QUERY =
   "(max-width: 640px), (max-width: 932px) and (max-height: 500px) and (orientation: landscape)";
@@ -70,6 +71,9 @@ function dismissChatComposerPickersOnEscape(event: KeyboardEvent): void {
     event.key !== "Escape" ||
     document.querySelector(".shell-nav[aria-modal='true']")
   ) {
+    return;
+  }
+  if (clearChatModelSearchOnEscape(event)) {
     return;
   }
   const pickers = openChatComposerPickers();
